requirements.txt
cached-property==2.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
confuse==2.0.1
cryptography==43.0.3
defusedxml==0.7.1
dnspython==2.7.0
exchangelib==5.5.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
msal==1.31.0
pywin32==308
numpy==2.1.3
oauthlib==3.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
six==1.16.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
